+++ src/day1_2021_sonic_sweep_part_2/Source.cpp
@@ -2,8 +2,8 @@
 #include <fstream>
 #include <iostream>
 #include <vector>
-int32_t CountIncreaseMeasurementsWindows(std::vector<int32_t> sonic_sweep) {
-    int32_t cnt = 0, sum_of_three1, sum_of_three2;
+uint64_t CountIncreaseMeasurementsWindows(std::vector<uint64_t> sonic_sweep) {
+    uint64_t cnt = 0, sum_of_three1, sum_of_three2;
     for (size_t i = 0; i < sonic_sweep.size() - 3; i++) {
         sum_of_three1 =
             sonic_sweep[i] + sonic_sweep[i + 1] + sonic_sweep[i + 2];
@@ -16,8 +16,8 @@ int32_t CountIncreaseMeasurementsWindows(std::vector<int32_t> sonic_sweep) {
     return cnt;
 }
 
-int32_t main() {
-    int32_t number;
+uint64_t main() {
+    uint64_t number;
     std::fstream input;
     // data.txt cotain input taken from
     // https://adventofcode.com/2021/day/1/input
